cdk/requirements.txt
aws-cdk-lib==2.150.0
aws-cdk.aws-lambda-python-alpha==2.150.0a0
constructs>=10.0.0,<11.0.0
datadog-cdk-constructs-v2
